Refine runtime import boundaries
