+++ afd_plugin/v1/worker/dbo.py
@@ -5,6 +5,8 @@
 from typing import Any
 
 import torch
+from vllm.utils.torch_utils import direct_register_custom_op
+from vllm.v1.worker.ubatching import dbo_enabled, dbo_yield
 
 _AFD_DBO_YIELD_OP_REGISTERED = False
 
@@ -30,8 +32,6 @@ def register_dbo_yield_custom_op() -> None:
 
     if _AFD_DBO_YIELD_OP_REGISTERED:
         return
-
-    from vllm.utils.torch_utils import direct_register_custom_op
 
     def afd_manual_dbo_yield_op(x: torch.Tensor) -> torch.Tensor:
         _yield_if_dbo_enabled()
@@ -72,8 +72,6 @@ def _yield_if_dbo_enabled() -> None:
     ):
         ascend_dbo_yield()
         return
-
-    from vllm.v1.worker.ubatching import dbo_enabled, dbo_yield
 
     if dbo_enabled():
         dbo_yield()
